requirements.txt
python-dotenv==0.10.3
python-mercuryapi==0.5.2
termcolor==1.1.0
flask==1.1.1
flask_cors==3.0.8